Drag+drop progress
Partially works now :)
